+++ marsworks/__init__.py
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-__version__ = "0.7.0"
+__version__ = "0.7.1"
 from marsworks.manifest import *
 from marsworks.origin import *
 from marsworks.photo import *